+++ AEWatchRenderManager/MainWindow.xaml.cs
@@ -18,8 +18,8 @@ using System.Windows.Controls;
 
 namespace AEWatchRenderManager
 {
-    // Date: Wed Apr 15 11:49:02 JST 2026
-    // Version: 1.16.14
+    // Date: Thu Apr 16 11:32:10 JST 2026
+    // Version: 1.16.15
     public partial class MainWindow : Window
     {
         public MainWindow()